+++ LotgdFormat/src/Nodes/Node.cs
@@ -17,6 +17,18 @@ internal readonly struct Node {
 		}
 	}
 
+	/// <summary>
+	/// Creates a new Text node.
+	/// </summary>
+	/// <param name="textStart">
+	/// Start of the text content in the input string.
+	/// </param>
+	/// <param name="textLength">
+	/// Length of the text cotent
+	/// </param>
+	/// <param name="IsUnsafe">
+	/// Whether the text is unsafe and needs to be HTML-Encoded when rendering.
+	/// </param>
 	public Node(int textStart, int textLength, bool IsUnsafe) {
 		this.Type = NodeType.Text;
 		this.TextStart = textStart;
@@ -26,6 +38,13 @@ internal readonly struct Node {
 
 	public Node(NodeType type, LotgdFormatCode code) {
 		this.Type = type;
+		this.Token = code.Token;
+		this._code = code;
+		this.Size = 0;
+	}
+
+	public Node(LotgdFormatCode code) {
+		this.Type = code._nodeType;
 		this.Token = code.Token;
 		this._code = code;
 		this.Size = 0;
@@ -52,10 +71,13 @@ internal readonly struct Node {
 						return " ";
 					}
 					case '\n': {
-						return "\n";
+						return "";
 					}
-					case > 'a' and < 'z':
+					case '"': {
+						return "&quot;";
+					}
 					case > '0' and < '9':
+					case > 'a' and < 'z':
 					case > 'A' and < 'Z': {
 						return text[0].ToString();
 					}